+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 =====
 
 * (feature) Improve rendering of admonitions.
+* (improvement) Improve styling of `comment()` and `done()`.
 
 
 1.3.0